pull out templating stuff
