+++ nextcloudappstore/core/static/assets/app/dom/TemplateEmpty.ts
